Rename run-specific session viewer components with run prefix
Disambiguate run session viewer components from the normal session
viewer by prefixing with "run": sessionViewer → runSessionViewer,
sessionViewerContainer → runSessionViewerContainer, etc.
