+++ app/modules/runs/components/runSessions.tsx
@@ -4,7 +4,7 @@ import type { Breadcrumb } from "~/modules/app/app.types";
 import Breadcrumbs from "~/modules/app/components/breadcrumbs";
 import type { Run, RunSession } from "~/modules/runs/runs.types";
 import SessionListSidebar from "~/modules/sessions/components/sessionListSidebar";
-import SessionViewerContainer from "~/modules/sessions/containers/sessionViewerContainer";
+import RunSessionViewerContainer from "~/modules/sessions/containers/runSessionViewerContainer";
 import type { SessionFile } from "~/modules/sessions/sessions.types";
 
 export default function RunSessions({
@@ -61,7 +61,7 @@ export default function RunSessions({
             <Spinner className="size-6" />
           </div>
         ) : (
-          <SessionViewerContainer
+          <RunSessionViewerContainer
             run={run}
             session={session}
             sessionFile={sessionFile}
